[Fix] Remimber 앱 > 잔디 공유 > Topic or message 선택 후 공유하기 시도 시 액션 없음 (JA-819) - 멀티 쉐어 업로드시에 멘션 관련 로직과 충돌되어 1:1메세지로 공유가 되지 않는 문제 해결

+++ app/src/main/java/com/tosslab/jandi/app/ui/share/multi/MultiShareFragment.java
@@ -258,7 +258,8 @@ public class MultiShareFragment extends Fragment implements MultiSharePresenter.
     @Override
     public void moveRoom(long teamId, long roomId) {
 
-        Completable.fromAction(() -> {})
+        Completable.fromAction(() -> {
+        })
                 .observeOn(AndroidSchedulers.mainThread())
                 .subscribe(() -> {
                     if (getActivity() == null) {
@@ -303,7 +304,7 @@ public class MultiShareFragment extends Fragment implements MultiSharePresenter.
 
     public void onEvent(ShareSelectRoomEvent event) {
         long roomId = event.getRoomId();
-        long roomType = event.getRoomType();
+        int roomType = event.getRoomType();
         multiSharePresenter.onSelectRoom(roomId, roomType);
 
     }
